requirements.txt
requests>=2.31.0
feedparser>=6.0.10
PyYAML>=6.0.1
urllib3>=2.0.0
